Replace ReservedRule slots with a separate vector
This commit changes the mechanism by which we disallow local names to be
reused by globals. Previously, local names were given a special slot in
RulesWithLocs::rules_[], of type ReservedRule. This commit changes it to
use a new, separate vector<Ident> reservedLocalNames_[].

The previous approach created too many special-cases for ReservedRule.
See the new testdata/ .oalex file for an example of what went wrong. We
essentially forgot to report an undefined variable because we expected
them all to be of type UnassignedRule. Using the same name both as a
global and a local, but never defining the global caused errors to
disappear.

Many other cases that needed to special-case ReservedRule can now be
simpler.

To accomplish this, we moved the registerLocations() call so that it is
used only for globals. Arguably, we should have already been doing this.