Delete registerLocations()
Apparently this function is merely a bandaid that we have been
redundantly maintaining for months.

This function was added in ae6fd56c to handle cases such as
testdata/rule-undefined-ident, where it helped report errors for rules
that were used but never defined. However, unknown to me, this became
redundant as a result of a cleanup commit a month later, 349222a5. This
commit was found through a git bisect.