Allow ExternParser to have a separate rule name
Previously, ExternParsers were called directly with their external name.
This makes it difficult to create aliases of the kind we want for
parameterized extern parsers.

For now, the Rule::nameOrNull() value is not really being used for
ExternParser. The next few rules will change this by introducing
forwarding parsers useful for parameterizing.