Implement `extern rule`
This is the first piece to be bootstrapped into oalex. Testing this is
slightly difficult since eval() didn't historically support extern
rules. This is why this commit also introduces the first oalexBuiltin.
It's a simple oalexBuiltinHello() just used for testing.

The new files frontend_pieces.{cpp,h} are generated from
frontend_pieces.oalex, but checked in to break build cycles.