Export line-skipping utility to lexer.h
Previously, lexer.cpp was internally using the singular form
skipBlankLine(). It had to be invoked in a loop. This commit exports the
plural form skipBlankLines().

Previously, this tedium was fine since frontend.cpp relied exclusively
on lexNextLine() which used this kind of a loop internally. Soon,
however, frontend.cpp will also use bootstrapped rules that will need to
also skip blank lines but will do its own lexing. Hence it is more
important now for this utility to be separated out.