Do not discard stack trace with rethrow
Rethrowing an exception with the parameterless `throw;` discards the
original exception stack trace. In this case, it replaced the trace with
something a lot less useful, with nothing but `oalex_main.cpp`
indicating the context of the error.

We cannot directly use `std::scope_exit` yet, since that's experimental.
So this commits makes a custom version of that.