Don't depend on stPos/enPos for parsing
The fields JsonLoc::{st,en}Pos were meant for reporting diagnostics.
Being not part of the functional requirements, they have so far not been
well-tested, which means it has remained buggy. Until we fix that, we
shouldn't be depending on these fields for parsing. This commit undoes
the one known example where we made the mistake of using it.