Fix codegen(ExternParser) with params
When ExternParser has parameters, they have to be provided as Parser
objects, not as function pointers. This changed in e9adc47e, so codegen
was broken since then. Tests hadn't caught this, because both the
codegen and tests were broken in a compatible way (in that both were
using function pointers). The problem only became apparent when we tried
testing codegen with one of the new parsers that used Parser objects
instead.