Clean up appendPatternRules() semantics
We have two overloads of this function with two slightly different
semantics. One is used when the `outputs:` stanza is available in a
`rule`, while the other is used when it isn't. It turns out that they
were slightly inconsistent. In particular, we had no way of ignoring a
`PartPattern`, where some parsed component is simply discarded in a
`rule`.