Remove Resetter from lexQuotedString()
We are not under memory pressure so severe that we will need to call
markUsed() all the time. We can reintroduce a coarser-grained mechanism
later if memory does become an issue.

In the meantime, this makes it complicated for us to separate out the
extraction of a string literal from the unescaping part. A separate
function for removing quotes and escapes from a string will be useful
for bootstrapping a little more.