Separate string-extraction from unquoting
lexQuotedString() was doing too much in one function. This commit
separates out the two steps into their own function. As an upshot, we
now have a version of this function that can work directly on a
StringLoc object instead of InputDiags. This can now be used on
the outputs from bootstrapped parsers.