Don't produce flattenable rules
Rules explicitly defined by the user should not be flattenable. This is
in contrast with component rules implicitly defined as a subpattern. It
turns out that we were using `ConcatFlatRule` where we should have had
something non-flattenable. This commit tries to automatically deduce the
right output template as we advertise doing.