Start the RuleExpr class hierarchy
This will be used for a new rule-list syntax that's "dumber", in that it
will not automatically insert SkipPoint like Pattern objects do.

This commit introduces a very simple single-string rule. Later commits
will expand this into an entire AST type like Pattern objects are. Most
of this commit is for setting up the unit-testing scaffolding. This
includes a way to test vector<unique_ptr<Rule>> for isomorphism.