Implement RuleExprMappedIdent
This change implements what will later be available is (var ~ /[a-z]+/)
syntax. It's a RuleExpr subexpression that specifies a field name.

Each dynamic_cast<> in compiler.cpp has been packaged up as a separate
method of a new RuleExprCompiler internal class (process() and
processMappedIdent()). The idea being, we will very likely need to
expand the list of branches with casts to other types that we will add
in the following commits.

assignRuleExpr() now needs to wrap the output in OutputTmpl{} to avoid
producing a flattenable output.

In tests, assertEqualJsonTmpl() currently supports only a subset of
JsonTmpl types since that's all we currently use in OutputTmpl.