Test if bootstrapped files need to change
This new test generates frontend_pieces.{h,cpp}, and checks if they are
still the same as before. If not, it flags an error and provides the
user with a command for regenerating it. This will prevent the kind of
errors that we saw in the previous commit.

It's not too eager in that it doesn't actually replace the source copy,
since we don't want to be crippled by some bug that we had just
introduced while coding.