Fix error-reporting on invalid identifiers
If a rule name started with a digit, e.g. `rule 0hello:`, oalex was
producing this puzzling error:

  Unexected 'rule', was expecting 'example' or 'rule'

It turns out that our `resemblesIdent()` function was too strict. It
needs to accept `0hello` as something that resembles an identifier, so
that a better error message can be provided later (e.g. "Expected
identifier").