Remove default comments
We were previously using C-style comments as the default for the user
language, since we needed to test Skippers. Now we can do that by adding
lexical options directly.

It'd be surprising to the user to see C-style comments being discarded
regardless of the language they are trying to define.