Return the full RuleSet from releaseRules()
Previously, this function was only returning the vector of rules,
leaving it to the caller to remember to initialize the other fields.
While this has been fine so far, we will soon have additional members of
RuleSet also initialized by RulesWithLocs. It will be easier if the
entire object is now set up in RulesWithLocs.