Don't hardcode oalexSkip in compiler.cpp
It's okay to use oalexSkip in frontend.cpp or lexer.cpp since they
process oalex source. But compiler.cpp deals with source in user
language, so they should not be hardcoded. Instead, it should be passed
in from the frontend, where it can be overridden when oalex acquires
such a capability.