Implement lexical stanza
This allows users to override comment patterns in specific rules. This
override applies both to comments in the pattern definition and the
input language.

Caveat: single-line comments, while they work in user inputs, don't seem
to be working in patterns. This seem to be an unrelated issue stemming
from how we trim out newlines from the end of patterns. Will be fixed in
a separate commit.