Use context_skipper() on user input
Previously, comments were not well-supported in examples or user input
when they appeared in the beginning or end of the input. This commit
utilizes the new context_skipper() to skip over these comments before
and after parsing the input.

This also happens to remove the last use of defaultSkip(), which can now
be deleted