Add method RulesWithLocs::defaultSkipper()
This commit doesn't change any functionality, just adds a new method.

First, this helps the deduplication heuristics in addSkipper() work a
bit better, by guaranteeing the default Skipper to be in the first
position (if it exists at all).

The other reason for this change is the commits that are coming next.
Right now, we do a poor job of supporting trailing or leading comments
in example inputs. To process them properly, we will augment Rule
objects with associated Skipper objects, so that they can be referenced
when testing examples.

This commit introduces the concept of a "default Skipper" to the
compiler, so that it can be used when constructing Rules. At least when
it is not overridden by the user.