Remove Skipper::withinLine() from lexer.cpp
This one was more involved than the other removals. oalexSkip now uses
the more appropriate newlines=ignore_blank, but that means a number of
changes in lexer.cpp, plus a minor simplification in frontend.cpp.

frontend.cpp doesn't need to use oalexSkip directly. It can call
skipBlankLines() instead.

In lexer.cpp:

  * The comment about forgetBefore() and acrossLines() is now stale.
    Removed.
  * lexSectionHeaderContents() no longer consumes the trailing newline,
    to adhere to the emerging convention of handling skippers in the
    rule that concatenates. It is now consumed in the caller instead.
  * Similarly, lexDashLine() is also much simpler, in that it doesn't
    do much about newlines around it.
  * The complexity has moved to lexSectionHeader(), where the unusual
    usage of keep_all can be localized.
  * lookaheadStart() uses ignore_all to effectively revert to the old
    behavior of oalexSkip that didn't use ignore_blank. This is needed
    since this function is used within BracketGroup parsing.
  * skipBlankLines() needs an extra explicit skip for the first newline,
    in case we are starting from the end of the previous line.
  * lexNextLine() optionally consumes a trailing newline, in keeping
    with the previous behavior of withinLine()
  * lexSourceLine() now properly uses keep_all, rather than relying on
    withinLine()

It is expected a lot of this mess can be cleaned up once we move to the
skip-on-concat convention more uniformly. It should now be easier once
we have migrated away from Skipper::withinLine().

Missed one use of withinLine()

Last one, for real.