Fix nullptr checks
Components in RuleExprConcat, RuleExprOptional, and RuleExprRepeat::part
must never be null. If parsing a child component produces nullptr, we
should discard them or return nullptr ourselves instead of adding
nullptr as a component in a valid object.

The only case where an aggregate RuleExpr object supports a null
component is RuleExprRepeat::glue. This should be checked for null value
before being dereferenced.