Fix blank-line handling bug
This commit fixes a bug for the following input:

   abc
   xyz

If we started skipping right after abc, with `newline = ignore_blank`,
it would often miss the newline right after the 'c'. It turns out that
the skipper needs context about where the line _really_ started for this
blank-line detection to work. Yet, there were many cases where we were
using skippers on substrings that lacked this necessary context.

This bug has to be fixed in three different places in the same commit,
since they conspired to hide the bug from the tests.

Fixing this bug revealed an "Ambiguous ellipsis" error in
rule-good-no-semicolon.oalex. This was fixed by adding a disambiguating
line in the test input.