Implement `words:` directive
This allows us to specify the character class for words in a pattern.

This commit uses all the refactoring from the series of previous
commits, by finally utilizing parseRegex() in a context where only
GluedString was available. This commit shows how we can now (explicitly)
convert a GluedString back to an InputDiags while preserving error
location and without creating a new buffer for inputs.