Use `InputPiece&` in parsing, not `Input&`
Most parsers do not actually care about the fact that the input is
coming from a stream, or the fact that we might have the capability to
forget parts of the input already processed. Most parsers only care
about (1) a string-like object, (2) whose parts can be mapped back to
the original inputs line and column number.

An abstract class is good enough for all of that. This change will help
us use existing parsers on substrings (e.g. GluedString) that have
already been extracted from a larger InputDiags. This abstract class
(InputPiece) is not tied to how we access the underlying input buffer,
or even how rowCol() ties back to the original location for error
reporting.

Still to do: `class Diag` still accepts an `InputPiece&`, since
`InputPiece` does not have a `rowCol()` method yet.