Revert "Move rowCol() from InputPiece to Input"
This reverts commit fd4bf14d3f45b4ded0c62828261d8ef27eb383aa.

The rowCol() method was originally removed from InputPiece because, at
the time, it required GluedString to retain an unowned pointer to Input
that risked dangling. We can regain the same functionality without any
Input* member, by tracking more location metadata internally.

Another reason this was because of a lack of convention. At the time we
had a mess of some parsers using InputDiags, some using GluedString,
some using InputDiagsRef (an old, leaky version of DiagsDest). Worse, we
could sometimes convert some of these types to another, but it was
unclear when. This was removed to try and change functions using
GluedString to them directly using InputDiags.

Today, we can reintroduce it since the conventions are much clearer, and
we are unlikely to regress into the mess we had. While we will
reintroduce _some_ conversions from GluedString back to InputDiags, they
will use explicit conversion functions and will be relatively much
rarer. Going forward, no parser will directly accept just a GluedString
in lieu of InputDiags.

  * If they are processing bytes, they will use InputDiags& and size_t&.
  * If they are working on processed tokens, they might use GluedString,
    but with additional DiagsDest explicitly provided.