Switch Diag back to InputPiece
We still need to change InputDiags, but that requires more commits.

This commit reverts the diags.* parts of
a38e881ada6afed57c5e50e964395aa7e1d3a615

The reason for this revert is explained in much greater detail in
c0f19db. This revert is undoing parts of the same original refactoring
effort.

The other part of a38e881 (pattern.cpp) might also get reverted later.
But if that happens, we will do it in a separate commit. Those are not
currently in the way of us using the parsers we want to use on
GluedString.