Remove Resetter
It was an interesting experiment that lasted for almost three years. The
Resetter was originally introduced way back in commit 761dd56. But it
turns out that composable parsers are fundamentally at odds with little
components that locally decide which part of the global input buffer can
be discarded or kept.

Over the years, the Resetter remained local to a few top-level functions
only. While we thought that is still useful enough to keep around, it
had indirect effects on our architecture. InputDiags was forced to keep
an Input type rather than an abstract InputPiece, just so memory
management was possible for these rare cases. This finally came to a
head when we were trying to change all parsers to use abstract types, to
improve error-location for example parsing and to be able to use
InputDiags in regex parsing. The Resetter type had to go.

We might still reintroduce Input::forgetBefore() calls in some chosen
parts of the frontend.cpp main loop. But that is for another day.