Remove InputDiags::markUsed()
This change was enabled by removal of all the Resetter objects in the
previous commit. It will help us when we finally transition over to
using InputPiece& in the InputDiags class, instead of Input. The removal
ensures that we don't accidentally regress back to using Input-specific
methods where we are not supposed to.