Start tracking input{Line,Col} in GluedString
Nobody uses these new fields yet.
They will get unit-tested in the next commit.

We want to start using general parsers on GluedString. These parsers
currently operate on InputDiags& with its Input member. If we can
consume both, we will be able to use existing functions to parse
indented inputs, and we will be able to use regex_io.h functions on
tokenized inputs.

This in turn requires GluedString to support rowCol(). When reporting
error messages, it needs to be able to map string positions back to
locations in the user-provided file. Currently, parsers do this by using
Input::rowCol(). This commit adds support for this in GluedString too.

The next commit actually adds the rowCol() method.