Move IndexRelation to cpp file
This type is an implementation detail, and do not have to be exposed to
other translation units. Many of the constructors and special member
functions now have to be defined in the .cpp file explicitly, as opposed
to being implicitly defined in the header. This is because their
implicit implementation also requires IndexRelation to be fully defined.

While it's more verbose, it's hopefully better for faster compile times.