Implement keyword tailcont
We can now suppress the extra newline expected at the end of
newline-sensitive patterns. The new keyword means that the tail of the
pattern continues on this line. Probably not the best keyword, but we
probably won't need it too often anyway.

A fair bit of complexity goes into classifying exactly what the error
is, and then trying to provide useful error messages to the user.