Add symtab parameter to assignRuleExpr()
Local rules will now be able to reference each other. So
assignRuleExpr() now recieves a map of local variable names, separate
from the names of rules in RulesWithLocs.

While testing this change, we found that it was useful to be able to
inspect StringRules, so we added a new case in ruleListDebugPrint() too.