RuleExprCompiler now supports a symbol table
For now, this commit allows the user of RuleExprCompiler to override a
subset of name lookups by providing a pre-resolved symbol table. Idents
are mapped to rule indices.

This commit does not _yet_ change any externally visible functionality,
since the new member symtab_ is always kept empty and since
RuleExprCompiler is an internal type anyway. The next commit will
populate this member.