Merge makePartPatterns
Previously, the caller was directly calling two overloads of
makePartPatterns(), one taking in a JsonTmpl and the other taking a
pattToRule vector. This commit merges the two overloads together, and
has the caller just that merged one.

This is so the newly merged function can be reused for other purposes,
such as checking errors about whether some parts are going unused, or
whether they appear in a pattern multiple times.

It looks like a merged function was the initial plan anyway. I'm
guessing I later kept it separate just for the sake of keeping functions
small.