Check local rules for multiple usage in pattern
We used to silently accept an error case that is fixed by this commit.
If the `where` stanza defined a rule that appeared multiple times in the
pattern, we were not previously catching that error. Instead, we were
silently producing invalid parsers. We _did_ catch multiple appearance
if the pattern part was defined in the `output` template.

The fix for this is to change `checkMultipleTmplParts()` to accept all
pattern part names, not just the ones that appear in the `output`. The
function is now renamed to `checkMultipleUsage()`, and its parameter
type is now different so it can accept the unified set of patterns.