Pull Ident collection out of checkMultipleUsage
Previously, checkMultipleUsageRecur() was a recursive function that
complained if a part was used multiple times in a pattern.

This commit splits this function into two separate ones.

  * One is patternCollectIdent(), that simply collects all parts used in
    a (compiled) Pattern tree, and returns a sorted vector<Ident>.
  * The other function is now called checkMultipleUsage(), that simply
    checks for duplicates in this vector.

This is helpful since, in a later commit, we will also use this
vector<ident> for other things:

  * It will be used for deducing the output template if that's absent.
  * It will be used to check for _unused_ local definitions (as opposed
    to the ones used multiple times).

Today, the deduced output template simply includes all local rules.
Which is fine, since they all need to be used in the pattern as well
today. But it won't work if we introduce helper local rules that won't
need to appear in the output.

The story is similar for the unused definition check. We will need a
version of this check that raises an error only if it is unused in the
pattern, and unused as a local helper rule.

For now, this commit doesn't change any functionality, aside from the
error message produced.