Allow helper-only local rules
Previously, all local rules were required to appear in the rule pattern.
But this check has now become too strict, since we can now have helper
local rules that are used by other local rules in the same scope.

This commit effectively moves where we perform the check for unused
pattern placeholders. Previously, this was being done in pattern.cpp,
when compiling a pattern. However, this part does not know which local
rule is used by other rules in the same scope.

So this commit performs the same check in compiler.cpp instead of
pattern.cpp. Here, we have all the information necessary to decide which
local rule is useful and which is not.