Fix local name lookup for RuleExprMappedIdent
Expressions like `(lhs ~ rhs)` were still looking up `rhs` in the set of
global rules, bypassing the local symbol table. This commit fixes this
behavior.

This change uncovered yet another bug, also fixed here.
ruleExprCollectIdent() was collecting the set of identifiers _produced_
in a RuleExpr, while it was being used in checkUnusedParts() for the set
of identifiers _used_ by rules.