Stop defining redundant local rules
Previously, we were being a little sloppy by defining a rule twice
sometimes. This happened when a name appeared both in the `outputs` and
in the `where` stanzas. It was mostly "harmless" since `mapToRule`
returns a vector: the later definition simply got ignored.

This commit cleans it up, by processing local definitions with the
`where` stanza only. This cleanup is useful now, since we are about to
make the definition of local rules slightly more complicated.