Fix local name lookup for simple cases
This commit deletes the function allocateRuleExpr() which had faulty
name resolution. It tried to cover for the functionality that is now
present in AliasRule by shortcutting some simple special cases. Not only
is this logic now redundant, it doesn't even attempt to look up local
rules.

While this does de-shortcut a few rules as a result, that can be left to
a later optimization phase if we ever have one. Right now, this allows
us to correct our name lookup logic.