Rename single-line rules to expression rules
They are no longer going to be single-line. We are adding `where`, and
`lexical` to them. Besides, it was always a little awkward to talk about
"non-single-line rules", since they weren't always multi-line either.

Going forward, we will refer to them as expression rules (vs. the other
kind that's a pattern rule). This naming is already visible in
compiler.h, where we have appendPatternRule() and appendExprRule(). This
commit propagates this convention in all other places I could find them.