Move p2rule_ into PatternToRulesCompiler
This will make the compiler class more self-contained. We will have one
fewer object whose lifetime we have to worry about. This was not a
problem so far, since objects of this class was always destroyed in the
same function that created it, compilePattern(). This will not remain
true, as we will be reusing the same compiler over multiple local
patterns. This usage is necessary for us to support double-quoted
strings in RuleExpr.