Generate RuleExprDquoted in the frontend
This commit also deletes a canary error that I had placed for myself,
back in 5ce183ac. I am now much less worried about accidentally
forgetting this, since the context skipper parameter is no longer
optiopnal in `defineIdent`. The overload that implicitly passes in
`Rule::removedContext` is now explicitly named `defineIdentForTest`.

In any case, for posterity, this error was to remind me that lexopts is
used in two separate places:

  1. when parsing a pattern string into a Pattern object.
  2. when calling defineIdent, if we are defining a global pattern
     that's a double-quoted string.