Pull compileLocalRules() out to a function
mapToRule() was doing two different tasks. They are now separate
functions. Yes, this does increase the caller's responsibility a little
bit for now. But this will help us implement compilation of
double-quoted strings in rules.

Parsing double-quoted strings will require us to keep more expensive
data structures that we will need to persist across calls of
assignRuleExpr(). E.g. the collection of Idents used in a pattern, or
the partPatterns objects used for tokenizing them. We can neither pass
these extra params to mapToRule() (because they have nothing to do with
mapping), nor can we recompute them each time assignRuleExpr() is called
(they are expensive). So this is the cleanest approach for now.