RuleExprDquoted compilation
This commit only adds parsing for double-quoted expressions, but can't
return the parsed parts in a syntax tree yet. Automatic
template-deduction depends on being able to collect all the output
identifiers, which will not be available until ruleExprCollectIdent()
works on RuleExprDquoted.

For now, we can have a very simple success or failure test.