Check for duplicates when collecting output Idents
It turns out that we were only checking for duplicates in
ruleExprMakeOutputTmpl(), which is only called if an output template is
_not_ given. If it is provided, no duplicate-checking was being done.

So now, we have moved the duplicate-output check from the place that
generated an output template to the place that collected output
identifiers. This is always done just before checking unused parts.

We took this opportunity to also return a vector properly instead of an
awkward output parameter.

Coincidentally, this also resolves the TODO about not repeating the
collection of output Idents.

The error message of some existing tests had to change from "Duplicate
identifier" or "Duplicate part" to just "Duplicate output", since we
are now reusing the same function in both pattern rules and expression
rules.