Don't require semicolons after `extern rule`
This resolves a TODO in `frontend_pieces.oalex`. The semicolon was a
hack from a time when oalex didn't support newline-handling very well.
This commit "resolves" the problem, however, not by making the grammar
sensitive to newlines. Instead, it made the parenthesis required, even
for empty parameter lists.

My original plan was to make the grammar sensitive to newlines. However,
it turned out that we were already using newlines in existing `extern
rule` statements in code and tests. This indicated that it was actually
desirable to allow linebreaks just about anywhere in `extern rule`.

I certainly did _not_ want `extern rule foo = bar()` to be immediately
followed by a `rule myrule` on the same line, without any intervening
newline. It turns out that was never a problem, since the frontend
lexer `skipBlankLines()` already disallowed them. This commit adds a
test case in `lexer_test.cpp` to make it official.

In theory, we could have still kept the parenthesis optional while
removing the semicolon. But that would have required unorthodox
approaches, such as making the `parseExternRule()` function consume a
trailing newline character. This is contrary to the oalex convention of
leaving surrounding whitespace to be processed by skippers in a parent
parser. And it would have resulted in a far more convoluted grammar,
where lexical newline overrides had to apply to very specific tokens.

Problems that remain: In this commit, we have convinced _ourselves_ that
an empty pair of parentheses `()` could be made mandatory. What about
user grammars? What if somebody _else_ wants a grammar where the
parentheses was optional? What if they also wanted the newlines to only
be an end-indicator in some specific cases?

It turns out that they can still do so, but using a more convoluted
arrangement.