Implement `lexical defaults:`
This feature allows the user to change defaults for all rules in a file.

To implement this feature, we ended up changing defaultSkipper() to
accept multiple calls. Previously, it could only be called once, and the
second call would always Bug() out. Now, it indicates errors with a -1
return value, and allows the default to be set multiple times as long as
no rules or additional skippers have yet been added to depend on it.