Expose defaultLexopts to parseOalexSource()
Several rule parsers were locally calling defaultLexopts() on their own.
Some were even hardcoding a copy of defaultLexopts() without calling it.
This makes it hard to change the default lexical options from one place.
This commit passes down the object as a parameter from the top level, so
that it is easy to override it from there.

This commit is purely a refactoring. Functional changes will be made in
a separate commit.

This change is necessary because we are about to implement a new
`lexical defaults:` feature that will allow the user to override the
default options for all rules in the file.

In order to accomplish this change, this commit adds a new `const
LexDirective&` parameter to a bunch of functions, and deletes
`defaultLexopts()`. It was becoming too tempting to call
`defaultLexopts()` when we should be using a parameter that can be
user-overridden, so it was better to delete the singleton instance to
remove the temptation.

To delete this function, we:

  * inlined it in `parseOalexSource()`, and
  * added a constructor to `struct RuleStanzas`

Adding this constructor means we will no longer be able to use a
designated initializer syntax like I had once wanted. But that's
probably okay, since we are calling it in only one place anyway.
We can revert back to explicit initialization if we end up needing the
designated initializer syntax later. It's an internal type.