Define the default RegexCharSet only once
Previously, this was being defined twice: once as part of
userRegexOpts, and later as part of defaultLexopts. This commit reuses
the second definition in both places. This is a refactoring commit only,
no functional changes.

This change will help when we introduce `lexical defaults:`, so that
these defaults can be overridden by the user in only one place.