Make RuleSet::regexOpts a vector
This doesn't change any functionality for now, since only one element is
allowed in the vector anyway.

It turns out that we are not currently using `lexical: words:` directive
in anything but Pattern tokenization. This was an oversight. Soon, we
will allow each rule to customize their notion of a "word" in the user
input.