Fix casts in needsName()
The `dynamic_cast` calls in needsName() were all wrong. The only reason
this bug was not previously caught is that needsName() is purely an
optimization. If it returns false too often, that just produces a few
extra rules in codegen.

The bug was introduced well over a year ago, in c35ed75 and 215a62c.
When we were cleaning up the old classes based on `std::variant` and
were moving to a class hierarchy (probably for faster build speeds).