Support inline calls only for `regexOptsIdx==0`
Previously, we were raising unimplemented errors for
`wp->regexOptsIdx != 0`, thinking we will implement the rest someday.
This commit makes this official: we don't plan on supporting those cases
at all. This is because the generated code will be too convoluted to
make these calls inline, along with the definition of a new
RegexCharSet. Instead, it makes more sense for such parsers to be in a
separate function.