Move regexOpts from releaseRules() to init
Previously, the compilation process in the frontend used to:
  * start by calling defaultSkipper(skip), and
  * finish by calling releaseRulesWith(regopts).

This commit changes the process, so we are now:
  * starting the compilation process by calling
    defaultLexopts({skip, regexOpts}), and
  * finishing by calling releaseRules().

The new methods are merely renamed versions of the older ones.
The new name describes their new roles better. During the compilation
process, the regexOpts_ is kept around in a new field of class
RulesWithLocs.

In addition to the frontend, the compilation process also had to be
updated in `compiler_test.cpp`.

As a result of this change, regexOpts (1) is now a size-one vector instead of
a single object, and (2) is available to RulesWithLocs much earlier in
the compilation process. It doesn't change any functionality right now,
but soon we will start to append to this vector for rules that specify a
`lexical: word:` directive.