Use `word:` directive in rule matching
While the `word:` directive was introduced months ago, it was only being
used in pattern tokenization. It was being completely ignored when
parsing user input (e.g. when testing a match for a regex word
boundary anchor). This commit fixes this oversight.

This commit takes in `lexopts.wordChars` and plumbs it all the way into
rules (WordPreserving{} and RegexRule{}). Previously, they used to
hardcode the `regexOptsIdx` to 0, in effect always using the default
word characters.

It also turns out that the method `appendRegexOrError()` and
`appendWordOrError()` aren't being used outside of compiler.cpp. Instead
of adding a new parameter to it in the header file, they have been made
statically linked functions internal to compiler.cpp.