Expose lexopts from RulesWithLocs with a getter
We previously kept two copies of defaultLexopts, neither of which were
canonical. One was disaggregated among the fields of RulesWithLocs,
while the other was stored locally in parseOalexSource(). The copies had
to be kept in sync each time they were modified, which was error-prone.
It was also confusing to read a local variable and a method in the `rl`
object having the same name.

This commit removes the local variable in parseOalexSource. Instead, we
call the getter on the `rl` object whenever needed. On the plus side, we
no longer have the awkward input-output parameter being  passed into
parseLexicalDefaults(). Calling `rl.defaultLexOpts(*lexopt)` is good
enough.

This commit is purely a cleanup.