Delete ConcatRule
It's been a long time since we started using ConcatFlatRule and
OutputTmpl instead of ConcatRule. The last place where the frontend
relied on ConcatRule was removed when we sunsetted the obsolete ':='
syntax in cc93c8f. The remaining code paths were not being exercised
anymore, so it's best to remove them now.

It will simplify our work of chaging return types for generated parsers.