Generate initial types for OutputTmpl
This commit simply generates the new struct definitions. Doesn't
actually use them in generated parsers yet. Moreover, it is only used
for OutputTmpl, and the fields are still hardcoded to be JsonTmpl types,
not the custom types produced by another rule.

LocPair definition is moved into input_view.h, so that it can be used
from all generated structs.