Change return type for OutputTmpl
Since it's an API change, it required a few different changes to happen
at once for everything to work out.

  * parserHeaders() now needs to output different headers based on
    whether it is for an OutputTmpl. So it needs to inspect the rule
    type.
  * The OutputTmpl now needs to return std::nullopt on error instead of
    JsonLoc::ErrorValue{}.
  * When passing a return value through a parent parser, we sometimes
    have to use a new toJsonLoc() utility. This converts an
    std::optional<ParsedMyRule> to JsonLoc. Some of these may be removed
    later as more rules use custom type.

frontend.cpp is also temporarily using toJsonLoc() just so we can
compile it using both the previous frontend_pieces.* as well as the one
generated here. It will be removed in the next commit.