Change return type of oalexBuiltins
This commit changes all the return types from JsonLoc to JsonLike. In
future, this will allow us to return more expressive structs rather than
pure JsonLoc.

This commit still has a number of hacks that need to be cleaned up.
We are doing this slowly to maintain the bootstrapping chain for
frontend_pieces.{h,cpp}, so that at each commit we can use the old
version of these files to generate the new one.