Change error representation in JsonLike
This commit makes all errors represented by an empty JsonLike::data_
field. Previously, it was possible to represent it using
JsonLoc::ErrorValue{} as well. This change allows us to easily implement
a JsonLike::operator bool() to check for errors.

In order to implement this change, we did a few things:

  * First, we added a new assertNotError() in the wrapper class so we
    don't accientally start returning JsonLoc::ErrorValue{}. This
    introduced a new jsonlike.cpp file, so the library is no longer just
    header-only.
  * JsonLoc ctor can now treat error values specially:
    - std::optional values are tested for std::nullopt.
    - JsonLoc is tested for hasErrorValue().
    - JsonLoc::ErrorValue is unconditionally discarded.

Finally, this change removes the need for toJsonLike(), since JsonLike
can now be directly constructed from std::optional. So we delete that
function here.