Utilities for interactive debugging
This commit introduces a temporary debugging aid,
debugNamedRuleFields(). It allowed me to visually check for a few oalex
files that we are indeed producing the correct fields.

Later, we can either

  * Remove this function, since we will be able to look at the generated
    output struct instead (when we can make those).
  * Make this a permanent fixture that we can automatically test from
    oalex files (e.g. `test rule fields myrule:`).

This commit also happens to start calling the new populateFlatFields()
from the frontend. They should now be populated for all flattenable
rules.